+++ src/main/java/com/mentorship/javaeats/repository/MenuItemRepository.java
@@ -1,7 +1,8 @@
 package com.mentorship.javaeats.repository;
 
-import com.mentorship.javaeats.model.MenuItem;
+
+import com.mentorship.javaeats.model.Entity.MenuItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuItemRepository extends JpaRepository<MenuItem, Integer> {
+public interface MenuItemRepository extends JpaRepository<MenuItem, Long> {
 }
