+++ src/main/java/com/mentorship/javaeats/repository/RestaurantDetailRepository.java
@@ -1,7 +1,7 @@
 package com.mentorship.javaeats.repository;
 
-import com.mentorship.javaeats.model.RestaurantDetail;
+import com.mentorship.javaeats.model.Entity.RestaurantDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantDetailRepository extends JpaRepository<RestaurantDetail, Integer> {
+public interface RestaurantDetailRepository extends JpaRepository<RestaurantDetail, Long> {
 }
